refactor: add Drive Error support
